+++ .gitignore
@@ -130,3 +130,7 @@ dmypy.json
 
 # Pycharm IDE Temp
 .idea/*.*
+
+# test_folders
+new_folder*
+test_folder
