+++ main.py
@@ -1,20 +1,14 @@
 import pyvips
 import os
+import sys
+from svgtojpg import *
 from svglib.svglib import svg2rlg
 from reportlab.graphics import renderPM
 
-class svgToimage() :
-    def __init__(self, path):
-        super().__init__()
-        self.path = path
-
-    def svgConvert(self):
-        image = pyvips.Image.new_from_file(self.path, dpi=300)
-        image.write_to_file("test.png") # fill none으로 되면 배경이 transparency.
-
-
+# python main.py 300
 if __name__ == '__main__':
-    inputPath = './test_files/66.svg'
+    inputPath = sys.argv[2]
+    inputDpi = sys.argv[3]
 
     converter = svgToimage(path=inputPath)
     converter.svgConvert()
